:rocket: завершил рефакторинг - вытащил сокеты в отдельный модуль, оставил комментарии с тудушками, small fixes
