--- client/api/index.js
@@ -1,4 +1,0 @@
-/**
-*
-* Здесь будет модуль, отвечающий за работу с сокетами
-*/
