+++ client/components/CanvasContainer/CanvasContainer.js
@@ -4,6 +4,7 @@ export default {
     name: 'canvas-container',
     store,
     computed: {
+        // TODO @dyadyaJora: разобраться
         // перерисовывать при любых изменениях в объекте Konva 
         // пока не знаю - косыль это или фича, потому что так 
         // будет перерисовываться на каждый чих
